Fix some compiler warnings by using "int" consistently for hash values.
git-svn-id: http://svn.boost.org/svn/boost/trunk@66072 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/regex/v4/basic_regex_creator.hpp
@@ -811,7 +811,7 @@ void basic_regex_creator<charT, traits>::fixup_recursions(re_syntax_base* state)
                // There may be more than one capture group with this hash, just do what Perl
                // does and recurse to the leftmost:
                //
-               idx = m_pdata->get_id(idx);
+               idx = m_pdata->get_id(static_cast<int>(idx));
             }
             while(p)
             {
